Refactor ladder_service to module
