+++ server/ladder_service/__init__.py
@@ -1,0 +1,7 @@
+from .game_name import game_name
+from .ladder_service import LadderService
+
+__all__ = (
+    "LadderService",
+    "game_name",
+)
